refactor: 이미지 패키지 이동 (#723)
