+++ backend/src/main/java/hanglog/trip/domain/Image.java
@@ -1,11 +1,10 @@
-package hanglog.image.domain;
+package hanglog.trip.domain;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 import hanglog.global.BaseEntity;
-import hanglog.trip.domain.Item;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
