+++ backend/src/test/java/hanglog/trip/service/ItemServiceTest.java
@@ -13,9 +13,9 @@ import hanglog.category.domain.repository.CategoryRepository;
 import hanglog.category.fixture.CategoryFixture;
 import hanglog.expense.domain.repository.ExpenseRepository;
 import hanglog.global.exception.BadRequestException;
-import hanglog.image.domain.repository.CustomImageRepository;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Item;
+import hanglog.trip.domain.repository.CustomImageRepository;
 import hanglog.trip.domain.repository.DayLogRepository;
 import hanglog.trip.domain.repository.ItemRepository;
 import hanglog.trip.domain.repository.PlaceRepository;
